added signal handling to get_line
